Predefine system table as global var (#15925)
把mo系统表和系统视图声明为全局变量，都引用自一处

Approved by: @daviszhen, @zhangxu19830126
